+++ src/cli/cliConfig.schema.js
@@ -1,0 +1,34 @@
+const cliConfigSchema = {
+  validFlags: [
+    {
+      name:
+      {
+        type: String,
+        required: true,
+      },
+      description: {
+        type: String,
+      },
+      shortCut: {
+        type: String,
+      },
+    },
+  ],
+  validOptions: [
+    {
+      name:
+      {
+        type: String,
+        required: true,
+      },
+      description: {
+        type: String,
+      },
+      shortCut: {
+        type: String,
+      },
+    },
+  ],
+};
+
+module.exports = { cliConfigSchema };