+++ src/execCmd.d.ts
@@ -1,4 +1,0 @@
-const { ChildProcessWithoutNullStreams } = require('child_process');
-
-export = execCmd;
-declare function execCmd(args: string[], cb: () => void): ChildProcessWithoutNullStreams;