+++ types/src/cli/index.d.ts
@@ -1,0 +1,3 @@
+export { parseArgv };
+import { parseArgv } from "./cli";
+//# sourceMappingURL=index.d.ts.map