+++ types/src/dmark/index.d.ts
@@ -1,0 +1,7 @@
+import { getConfig } from "./dmark";
+import { getLabels } from "./dmark";
+import { getStacks } from "./dmark";
+import { getStages } from "./dmark";
+import { executeCommand } from "./dmark";
+export { getConfig, getLabels, getStacks, getStages, executeCommand };
+//# sourceMappingURL=index.d.ts.map