+++ types/src/execCmd/execCmd.d.ts
@@ -1,0 +1,8 @@
+/**
+ * Execute an asynchronous cli command as sub process.
+ *
+ * @param {string[]} args An array of cli args.
+ * @returns {import("promisify-child-process").ChildProcessPromise | null} The cli results or null.
+ */
+export function execCmd(args: string[]): import("promisify-child-process").ChildProcessPromise | null;
+//# sourceMappingURL=execCmd.d.ts.map