+++ types/src/execCmd/execCmd.spec.d.ts
@@ -1,0 +1,2 @@
+export {};
+//# sourceMappingURL=execCmd.spec.d.ts.map