+++ types/src/utils/utils.d.ts
@@ -1,0 +1,8 @@
+export function runQueue(tasks: any): Promise<void>;
+export class runQueue {
+    constructor(tasks: any);
+    current: number;
+}
+export function cmdTask(next: any, args: any): Promise<void>;
+export function schemaValidation(obj: any, schemaObj: any, schemaName?: string): void;
+//# sourceMappingURL=utils.d.ts.map