+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
-## [1.0.0] - 2022-09-10
-### Feat
-- TODO
+## [1.0.0] - 2022-10-10
+### Added
+- Dmark binary entrypoint.
+- Dmark config schema validator.
+- Custom CLI helper.
+- Lint using ESLint.
+- Initial unit tests with Jest.
+- Types generation from JSDoc comments using tsc.